+++ Source/ViewControllerPresentationSpy/AlertVerifier.swift
@@ -3,15 +3,18 @@
 
 import UIKit
 
-/*!
- * @abstract Captures presented UIAlertControllers.
- * @discussion Instantiate an AlertVerifier before the execution phase of the test. Then invoke the
- * code to create and present your alert. Information about the alert is then available through the
- * AlertVerifier.
+/**
+    Captures presented UIAlertControllers.
+ 
+    Instantiate an AlertVerifier before the execution phase of the test. Then invoke the code to
+    create and present your alert. Information about the alert is then available through the
+    AlertVerifier.
  */
 @objc(QCOAlertVerifier)
 public class AlertVerifier: NSObject {
+    /// Number of times present(_:animated:completion:) was called.
     @objc public var presentedCount = 0
+    
     @objc public var presentingViewController: UIViewController?
     @objc public var animated: Bool = false
     @objc public var title: String?
@@ -21,13 +24,18 @@ public class AlertVerifier: NSObject {
     @objc public var preferredAction: UIAlertAction?
     @objc public var popover: UIPopoverPresentationController?
     @objc public var textFields: [UITextField]?
+
+    /// Production code completion handler passed to present(_:animated:completion:).
     @objc public var capturedCompletion: (() -> Void)?
+
+    /// Test code can provide its own completion handler to fulfill XCTestExpectations.
     @objc public var testCompletion: (() -> Void)?
 
-    /*!
-     * @abstract Initializes a newly allocated verifier.
-     * @discussion Instantiating an AlertVerifier swizzles UIViewController and UIAlertController.
-     * They remain swizzled until the AlertVerifier is deallocated.
+    /**
+        Initializes a newly allocated verifier.
+     
+        Instantiating an AlertVerifier swizzles UIViewController, UIAlertController, and
+        UIAlertAction. They remain swizzled until the AlertVerifier is deallocated.
      */
     @objc public override init() {
         super.init()
@@ -70,10 +78,10 @@ public class AlertVerifier: NSObject {
         }
     }
 
-    /*!
-     * @abstract Executes the action for the button with the specified title.
-     * @discussion Throws an exception (or returns an error in ObjC) if no button with that title is
-     * found.
+    /**
+        Executes the action for the button with the specified title.
+        
+        Throws an exception (or returns an error in ObjC) if no button with given title is found.
      */
     @objc(executeActionForButton:andReturnError:)
     public func executeAction(forButton title: String) throws {